Add error state integration tests

+++ test/integration/services/bmiCalculator.test.js
@@ -1,5 +1,7 @@
 const { expect } = require('chai');
 const fs = require('fs');
+const proxyquire = require('proxyquire');
+const { Readable } = require('stream');
 const { bmiCalculatorService } = require('../../../src/container');
 const { ranges: bmiRanges, calculateBMIKgMetersSquared, getBMICategory } = require('../../utils/bmi');
 
@@ -10,7 +12,9 @@ describe('bmiCalculatorService', () => {
 
   afterEach(async () => {
     // clean up test output file
-    await fs.unlinkSync(outputFile);
+    if (fs.existsSync(outputFile)) {
+      await fs.unlinkSync(outputFile);
+    }
   });
 
   it('calculates BMI data of the correct input file and writes it to an output file', async () => {
@@ -36,5 +40,91 @@ describe('bmiCalculatorService', () => {
         )
       ].healthRisk,
     }));
+  });
+
+  describe('when an error occurs reading the data', () => {
+    const { bmiCalculatorService: badBmiCalculatorService } = proxyquire('../../../src/container', {
+      fs: {
+        createReadStream: () => { throw new Error(); },
+      },
+    });
+
+    before(() => {
+      process.env.PORT = 1234;
+    });
+
+    after(() => {
+      process.env.PORT = '';
+    });
+
+    it('rejects with an error', () => {
+      expect(badBmiCalculatorService.calculate(outputFile)).to
+        .eventually.be.rejectedWith(Error, 'Unable to process BMI data');
+    });
+  });
+
+  describe('when an error occurs writing the data', () => {
+    const { bmiCalculatorService: badBmiCalculatorService } = proxyquire('../../../src/container', {
+      fs: {
+        createWriteStream: () => { throw new Error(); },
+      },
+    });
+
+    before(() => {
+      process.env.PORT = 1234;
+    });
+
+    after(() => {
+      process.env.PORT = '';
+    });
+
+    it('rejects with an error', () => {
+      expect(badBmiCalculatorService.calculate(outputFile)).to
+        .eventually.be.rejectedWith(Error, 'Unable to process BMI data');
+    });
+  });
+
+  describe('when the data contains bad data', () => {
+    const badData = '{ invalid }';
+    const otherData = [{
+      Gender: 'Male',
+      HeightCm: 171,
+      WeightKg: 96,
+    },
+    {
+      Gender: 'Male',
+      HeightCm: 161,
+      WeightKg: 85,
+    }];
+
+    const { bmiCalculatorService: badBmiCalculatorService } = proxyquire('../../../src/container', {
+      fs: {
+        createReadStream: () => Readable.from([
+          JSON.stringify(otherData.slice(0, otherData.length / 2)),
+          badData,
+          JSON.stringify(otherData.slice(otherData.length / 2)),
+        ]),
+      },
+    });
+
+    it('continues processing', async () => {
+      await badBmiCalculatorService.calculate(outputFile);
+
+      const outputData = fs.readFileSync(outputFile, { encoding: 'utf8' });
+
+      expect(() => JSON.parse(outputData)).not.to.throw();
+      JSON.parse(outputData).forEach((person, i) => expect(person).to.eql({
+        ...otherData[i],
+        BMI: otherData[i].WeightKg / (otherData[i].HeightCm / 100),
+        BMICategory: getBMICategory(
+          calculateBMIKgMetersSquared(otherData[i].WeightKg, otherData[i].HeightCm),
+        ),
+        HealthRisk: bmiRanges[
+          getBMICategory(
+            calculateBMIKgMetersSquared(otherData[i].WeightKg, otherData[i].HeightCm),
+          )
+        ].healthRisk,
+      }));
+    });
   });
 });
